requirements.txt
numpy
torch
torchvision
Pillow
osmnx
geopandas
rasterio
shapely
tqdm
matplotlib
scikit-image